+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-06T11:09:54.594Z",
+    "2026-02-06T13:25:31.849Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {